requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
dj-database-url==3.0.1
Django==5.1.3
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==23.0.0
h11==0.14.0
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.2
tzdata==2024.2
uvicorn==0.32.1
whitenoise==6.8.2
